+++ test/instagram.test.ts
@@ -1,16 +1,33 @@
 import { test, expect } from "vitest"
-import { getFollowerCount } from "../src"
+import { getFollowerCount, getIgSessionId } from "../src"
 
 test("instagram", async () => {
+  const sessionId = await getIgSessionId(
+    "dummy_ig_user_egoist",
+    import.meta.env.VITE_IG_PASSWORD,
+  )
   const count = await getFollowerCount({
     type: "instagram",
     username: "hello",
-    auth: {
-      username: "dummy_ig_user_egoist",
-      password: import.meta.env.VITE_IG_PASSWORD,
-    },
+    sessionId,
   })
 
   console.log(count)
   expect(count).toBeGreaterThan(0)
+})
+
+test("instagram user not found", async () => {
+  try {
+    const sessionId = await getIgSessionId(
+      "dummy_ig_user_egoist",
+      import.meta.env.VITE_IG_PASSWORD,
+    )
+    await getFollowerCount({
+      type: "instagram",
+      username: "hello_xx88aa",
+      sessionId,
+    })
+  } catch (error: any) {
+    expect(error.message).toContain("user not found")
+  }
 })